Jenkinsfile: set default for params
Builds are configured (including parameters) from the Jenkinsfile.
This means the very first build doesn't have params and therefore
they don't get their defaults set either.

This is only an issue on the very first build of a job but the
organisation plugin adds PR's as new jobs meaning the first build
of every PR fails.

Hence check the param even exists before using it and if not use
it's default.

Signed-off-by: Ian Pozella <Ian.Pozella@imgtec.com>